+++ src/Import.php
@@ -8,7 +8,7 @@
 * @package      PHP Library
 * @subpackage   phplibrary
 * @category     Files
-* @author       Ivan Skokić <iskokic@zgmail.com>
+* @author       Ivan Skokić <iskokic@gmail.com>
 */
 namespace phplibrary;
 
@@ -18,6 +18,8 @@ use PhpOffice\PhpSpreadsheet\IOFactory as IOFactory;
 * Import data from file using customisation class of PHPOffice/PhpSpreadsheet
 */
 class Import {
+    
+    // -------------------------------------------------------------------------
 
     /**
     * Available types
@@ -56,7 +58,8 @@ class Import {
         if (file_exists($file_path))
         {
             $file_extension = strtolower(pathinfo($file_path, PATHINFO_EXTENSION));
-            if(in_array($file_extension, self::$allowed_types))
+            
+            if (in_array($file_extension, self::$allowed_types))
             {
                 $spreadsheet = IOFactory::load($file_path);
                 $sheetData   = $spreadsheet->getActiveSheet()->toArray(NULL, TRUE, TRUE, TRUE);
@@ -69,5 +72,4 @@ class Import {
     }
 
     // -------------------------------------------------------------------------
-
 }