+++ changelog.md
@@ -1,10 +1,11 @@
 Since php-library v1.4.0 (10-Mar-2018)
 =======
 
-Removed
+Modified
 ----------------
 
 * Date_Time_Format: protected validate method removed
+* User: attributes $image_location and $image_default set as public
 
 Since php-library v1.3.0 (06-Feb-2018)
 =======